back button fix
You need to use "condition" to make different "onSomething" actions,
otherwise it won't work properly.

I don't know what 20 was used for, but if it needs to be reinstated,
then make a different back button visible only in this cases or better
yet make a
 <onup condition="windowisvisible(needed window)">20 </onup>

something like that. (check syntax) :)